% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{plot.matrix}
\alias{plot.matrix}
\title{Presents matrix as 3D surface and/or a heatmap}
\usage{
\method{plot}{matrix}(x, labs = c("X", "Y", "Z"), types = c("2D", "3D"),
  color.palette = c("default", "special"), ...)
}
\arguments{
\item{x}{Matrix to plot.}

\item{labs}{Vector of labels for X, Y and Z axes.}

\item{types}{Vector of plot types to plot. Possible types are \code{"2D"} and \code{"3D"}. Default value for the parameter is \code{c("3D", "2D")}.}

\item{color.palette}{Character string \code{"default"} or \code{"special"} or a function accepting one argument and returning a color palette
(for example \code{rainbow} from \code{grDevices} package).}

\item{...}{Other parameters. They are currently ignored.}
}
\description{
Presents matrix as 3D surface and/or a heatmap
}
\examples{
\dontrun{

plot(matrix(rnorm(100),10,10))
plot(matrix(1:100,10,10), c("Dimension 1", "Dimension 2", "Value"))

library(demography)
m <- log(fr.mort$rate$female[1:30, 150:160])
plot(m, type = "2D", color.palette = "default")
plot(m, type = "3D", color.palette = "special")
plot(m, color.palette = rainbow)
}
}
\author{
Alexander Dokumentov
}

