% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{plot.smAPC}
\alias{plot.smAPC}
\title{Presents demographic data using 3D surface and/or a heatmap}
\usage{
\method{plot}{smAPC}(x, component = c("all", "surface", "period", "cohort",
  "residuals", "original"), labs = c("Age", "Time", NA), types = c("3D",
  "2D"), color.palette = c("default", "special"), ...)
}
\arguments{
\item{x}{Result of smoothing (object of class \code{smAPC}).}

\item{component}{"smooth", "period", "cohort", "residuals" or "original"}

\item{labs}{Vector of labels for X, Y and Z axes.}

\item{types}{Vector of plot types to plot. Possible types are \code{"2D"} and \code{"3D"}. Default value for the parameter is \code{c("3D", "2D")}.}

\item{color.palette}{Character string \code{"default"} or \code{"special"} or a function accepting one argument and returning a color palette
(for example \code{rainbow}).}

\item{...}{Other parameters. They are currently ignored.}
}
\description{
Presents demographic data using 3D surface and/or a heatmap
}
\examples{
\dontrun{

library(demography)
m <- log(fr.mort$rate$female[1:30, 150:160])
sm <- autoSmoothAPC(m)

plot(sm)
plot(sm, "surface")
plot(sm, "cohort")
plot(sm, "period")
plot(sm, "residuals")
plot(sm, "original")

}
}
\author{
Alexander Dokumentov
}

