% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.sm2D}
\alias{print.sm2D}
\title{Prints the result of smoothing (object of class \code{sm2D}).}
\usage{
\method{print}{sm2D}(x, ...)
}
\arguments{
\item{x}{Result of smoothing (object of class \code{sm2D}).}

\item{...}{Other parameters. They are currently ignored.}
}
\description{
Prints the result of smoothing (object of class \code{sm2D}).
}
\examples{
\dontrun{

library(demography)
m = log(fr.mort$rate$female[1:30, 150:160])
sm = autoDemogSmooth(m)
sm

}
}
\author{
Alexander Dokumentov
}

